+++ packages/article-list/__tests__/mocks.js
@@ -1,5 +1,6 @@
 // eslint-disable-next-line global-require
 jest.mock("@times-components/ad", () => require("./ad-mock"));
+jest.mock("@times-components/article-summary", () => "ArticleSummary");
 jest.mock("@times-components/button", () => "Button");
 jest.mock("@times-components/card", () => "Card");
 jest.mock("@times-components/image", () => "Image");