+++ packages/card/src/card.android.js
@@ -1,0 +1,3 @@
+import CardContent from "./card-content";
+
+export default CardContent;