+++ packages/pages/src/article.js
@@ -58,6 +58,7 @@ const ArticleDetailsPage = ({
             onTopicPress={(event, { slug }) => onTopicPress(slug)}
             onTwitterLinkPress={(_, { url }) => onLinkPress(url)}
             onVideoPress={(event, info) => onVideoPress(info)}
+            pageSection={pageSection}
             refetch={refetch}
           />
         </Context.Provider>