+++ packages/card/CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.21"></a>
+## [3.0.21](https://github.com/newsuk/times-components/compare/@times-components/card@3.0.20...@times-components/card@3.0.21) (2018-08-24)
+
+**Note:** Version bump only for package @times-components/card
+
+
+
+
+
+<a name="3.0.20"></a>
+## [3.0.20](https://github.com/newsuk/times-components/compare/@times-components/card@3.0.19...@times-components/card@3.0.20) (2018-08-24)
+
+
+### Bug Fixes
+
+* refactor native image loading animation ([#1243](https://github.com/newsuk/times-components/issues/1243)) ([1195553](https://github.com/newsuk/times-components/commit/1195553))
+
+
+
+
+
+<a name="3.0.19"></a>
+## [3.0.19](https://github.com/newsuk/times-components/compare/@times-components/card@3.0.18...@times-components/card@3.0.19) (2018-08-23)
+
+**Note:** Version bump only for package @times-components/card
+
+
+
+
+
+<a name="3.0.18"></a>
+## [3.0.18](https://github.com/newsuk/times-components/compare/@times-components/card@3.0.17...@times-components/card@3.0.18) (2018-08-23)
+
+**Note:** Version bump only for package @times-components/card
+
+
+
+
+
 <a name="3.0.17"></a>
 ## [3.0.17](https://github.com/newsuk/times-components/compare/@times-components/card@3.0.16...@times-components/card@3.0.17) (2018-08-22)