+++ packages/markup/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.0.15"></a>
+## [3.0.15](https://github.com/newsuk/times-components/compare/@times-components/markup@3.0.14...@times-components/markup@3.0.15) (2018-08-24)
+
+**Note:** Version bump only for package @times-components/markup
+
+
+
+
+
+<a name="3.0.14"></a>
+## [3.0.14](https://github.com/newsuk/times-components/compare/@times-components/markup@3.0.13...@times-components/markup@3.0.14) (2018-08-23)
+
+**Note:** Version bump only for package @times-components/markup
+
+
+
+
+
+<a name="3.0.13"></a>
+## [3.0.13](https://github.com/newsuk/times-components/compare/@times-components/markup@3.0.12...@times-components/markup@3.0.13) (2018-08-23)
+
+**Note:** Version bump only for package @times-components/markup
+
+
+
+
+
 <a name="3.0.12"></a>
 ## [3.0.12](https://github.com/newsuk/times-components/compare/@times-components/markup@3.0.11...@times-components/markup@3.0.12) (2018-08-22)