+++ packages/tracking/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.13"></a>
+## [2.1.13](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.1.12...@times-components/tracking@2.1.13) (2018-08-24)
+
+**Note:** Version bump only for package @times-components/tracking
+
+
+
+
+
+<a name="2.1.12"></a>
+## [2.1.12](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.1.11...@times-components/tracking@2.1.12) (2018-08-24)
+
+**Note:** Version bump only for package @times-components/tracking
+
+
+
+
+
+<a name="2.1.11"></a>
+## [2.1.11](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.1.10...@times-components/tracking@2.1.11) (2018-08-23)
+
+**Note:** Version bump only for package @times-components/tracking
+
+
+
+
+
 <a name="2.1.10"></a>
 ## [2.1.10](https://github.com/newsuk/times-components/compare/@times-components/tracking@2.1.9...@times-components/tracking@2.1.10) (2018-08-22)